Migrate policy definition resource and data source test
